Update README.md
Watch live TV channels from across the world as the playlist gets updated on a daily basis, as well as throughout the day the number of channels increase. It will find new and working links and add them to the playlist automatically for you to enjoy. https://www.iptvfree.tv provides dynamic content with a static URL, thus making it possible to use the same URL for any device on the market.